02_Chat_With_PDF/requirements.txt
langchain_classic 
langchain_community 
langchain_aws 
PyPDF2 
tiktoken 
faiss-cpu
streamlit
python-dotenv